use WarningMessage class
