+++ lib/eventum/class.mail_helper.php
@@ -145,84 +145,6 @@ class Mail_Helper
     }
 
     /**
-     * Removes the warning message contained in a message, so that certain users
-     * don't receive that extra information as it may not be relevant to them.
-     *
-     * @param   string $str The body of the email
-     * @return  string The body of the email, without the warning message
-     */
-    public static function stripWarningMessage($str)
-    {
-        $str = str_replace(self::getWarningMessage('allowed'), '', $str);
-        $str = str_replace(self::getWarningMessage('blocked'), '', $str);
-
-        return $str;
-    }
-
-    /**
-     * Returns the warning message that needs to be added to the top of routed
-     * issue emails to alert the recipient that he can (or not) send emails to
-     * the issue notification list.
-     *
-     * @param   string $type Whether the warning message is of an allowed recipient or not
-     * @return  string The warning message
-     */
-    public static function getWarningMessage($type)
-    {
-        if ($type == 'allowed') {
-            $str = ev_gettext('ADVISORY: Your reply will be sent to the notification list.');
-        } else {
-            $str = ev_gettext('WARNING: If replying, add yourself to Authorized Repliers list first.');
-        }
-
-        return $str;
-    }
-
-    /**
-     * Method used to add a customized warning message to the body
-     * of outgoing emails.
-     *
-     * @param   int $issue_id The issue ID
-     * @param   string $to The recipient of the message
-     * @param   string $body The body of the message
-     * @param   array $headers The headers of the message
-     * @return  string The body of the message with the warning message, if appropriate
-     */
-    public static function addWarningMessage($issue_id, $to, $body, $headers)
-    {
-        $setup = Setup::get();
-        $enabled = $setup['email_routing']['status'] == 'enabled' && $setup['email_routing']['warning']['status'] == 'enabled';
-        if (!$enabled) {
-            return $body;
-        }
-
-        // check if the recipient can send emails to the customer
-        $recipient_email = self::getEmailAddress($to);
-        $recipient_usr_id = User::getUserIDByEmail($recipient_email);
-        // don't add the warning message if the recipient is an unknown email address
-        if (empty($recipient_usr_id)) {
-            return $body;
-        }
-
-        // don't add anything if the recipient is a known customer contact
-        $recipient_role_id = User::getRoleByUser($recipient_usr_id, Issue::getProjectID($issue_id));
-        if ($recipient_role_id == User::ROLE_CUSTOMER) {
-            return $body;
-        }
-
-        if (!Support::isAllowedToEmail($issue_id, $recipient_email)) {
-            $warning = self::getWarningMessage('blocked');
-        } else {
-            $warning = self::getWarningMessage('allowed');
-        }
-        if (@$headers['Content-Transfer-Encoding'] == 'base64') {
-            return base64_encode($warning . "\n\n" . trim(base64_decode($body)));
-        }
-
-        return $warning . "\n\n" . $body;
-    }
-
-    /**
      * Method used to save a copy of the given email to a configurable address.
      *
      * @param MailMessage $mail the email to save
