+++ tests/Mail/WarningMessageTest.php
@@ -14,8 +14,11 @@
 namespace Eventum\Test\Mail;
 
 use Eventum\Mail\Helper\WarningMessage;
+use Eventum\Mail\MailBuilder;
 use Eventum\Mail\MailMessage;
 use Eventum\Test\TestCase;
+use Mail_Helper;
+use Setup;
 use Zend\Mail\Exception\InvalidArgumentException;
 
 class WarningMessageTest extends TestCase
@@ -28,6 +31,26 @@ class WarningMessageTest extends TestCase
         $mail = MailMessage::createFromString("X-Foo: 1\r\n\r\nHello. Bääm");
 
         $this->runAddAndRemoveTests($mail, $issue_id, $email);
+    }
+
+    public function testMailBuilderText()
+    {
+        $this->markTestSkipped('not yet');
+
+        $issue_id = 1;
+        $recipient = $from = 'root@localhost';
+        $body = $subject = 'Hello. Bääm';
+
+        $builder = new MailBuilder();
+        $builder
+            ->addTextPart($body)
+            ->getMessage()
+            ->setSubject($subject)
+            ->setTo($recipient)
+            ->setFrom($from);
+
+        $mail = $builder->toMailMessage();
+        $this->runAddAndRemoveTests($mail, $issue_id, $recipient);
     }
 
     /**
@@ -54,6 +77,29 @@ class WarningMessageTest extends TestCase
         $wm->remove();
         $raw3 = $mail->getRawContent();
         $this->assertEquals($raw1, $raw3);
+    }
+
+    public function testAddWarningMessage()
+    {
+        $this->markTestSkipped('requires MailMessage port');
+        $setup = Setup::get();
+        $setup['email_routing']['status'] = 'enabled';
+        $setup['email_routing']['warning']['status'] = 'enabled';
+
+        $issue_id = 1;
+        $recipient = 'admin@example.com';
+        $body = 'here be dragons';
+
+        // add the warning message to the current message' body, if needed
+        $m = MailMessage::createFromHeaderBody([], $body);
+        Mail_Helper::addWarningMessage($issue_id, $recipient, $m);
+        $fixed_body = $m->getContent();
+        $this->assertContains($body, $fixed_body);
+
+        $m = MailMessage::createFromFile(__DIR__ . '/data/attachment-bug.txt');
+        Mail_Helper::addWarningMessage($issue_id, $recipient, $m);
+        $fixed_body = $m->getContent();
+        $this->assertContains('Your reply will be sent to the notification list', $fixed_body);
     }
 
     private function runAddAndRemoveTests(MailMessage $mail, $issue_id, $email)
